+++ src/ralph_assets/models_assets.py
@@ -42,10 +42,16 @@ from django.template import Context, Template
 from django.utils.translation import ugettext_lazy as _
 
 from ralph.business.models import Venture
-from ralph.discovery.models_device import Device, DeviceType
+from ralph.discovery.models_device import (
+    Device,
+    DeviceEnvironment,
+    DeviceType,
+    ServiceCatalog,
+)
 from ralph.discovery.models_util import SavingUser
 from ralph_assets.history.models import HistoryMixin
 from ralph_assets.history.utils import field_changes
+from ralph.util.models import SyncFieldMixin
 from ralph_assets.models_util import WithForm
 from ralph_assets.utils import iso2_to_iso3
 
@@ -406,6 +412,7 @@ class Asset(
     SavingUser,
     SoftDeletable,
     WithForm,
+    SyncFieldMixin,
 ):
     '''
     Asset model contain fields with basic information about single asset
@@ -543,6 +550,18 @@ class Asset(
         help_text=HOSTNAME_FIELD_HELP_TIP,
     )
     required_support = models.BooleanField(default=False)
+    service = models.ForeignKey(
+        ServiceCatalog,
+        default=None,
+        null=True,
+        on_delete=models.PROTECT,
+    )
+    device_environment = models.ForeignKey(
+        DeviceEnvironment,
+        default=None,
+        null=True,
+        on_delete=models.PROTECT,
+    )
 
     def __unicode__(self):
         return "{} - {} - {}".format(self.model, self.sn, self.barcode)
@@ -557,13 +576,9 @@ class Asset(
 
     @property
     def venture(self):
-        if not self.device_info or not self.device_info.ralph_device_id:
-            return None
         try:
-            return Device.objects.get(
-                pk=self.device_info.ralph_device_id,
-            ).venture
-        except Device.DoesNotExist:
+            self.get_ralph_device().venture
+        except AttributeError:
             return None
 
     @property
@@ -625,8 +640,26 @@ class Asset(
                 if different_country:
                     self.generate_hostname(commit, template_vars)
 
-    def save(self, commit=True, *args, **kwargs):
+    def get_ralph_device(self):
+        if not self.device_info or not self.device_info.ralph_device_id:
+            return None
+        try:
+            return Device.objects.get(
+                pk=self.device_info.ralph_device_id,
+            )
+        except Device.DoesNotExist:
+            return None
+
+    def get_synced_objs_and_fields(self):
+        # Implementation of the abstract method from SyncFieldMixin.
+        fields = ['service', 'device_environment']
+        obj = self.get_ralph_device()
+        return [(obj, fields)] if obj else []
+
+    def save(self, commit=True, sync=True, *args, **kwargs):
         _replace_empty_with_none(self, ['source', 'hostname'])
+        if sync:
+            SyncFieldMixin.save(self, *args, **kwargs)
         instance = super(Asset, self).save(commit=commit, *args, **kwargs)
         return instance